requirements.txt

torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.24.0

pandas>=2.0.0
scikit-learn>=1.3.0 

matplotlib>=3.7.0
